+++ TradeItIosTicketSDK2/TradeItTransactionsTableViewManager.swift
@@ -137,7 +137,7 @@ class TransactionHistoryResultPresenter {
         self.transactions = transactionHistoryResult.transactionHistoryDetailsList?.sorted{ ($0.date ?? "01/01/1970") > ($1.date ?? "01/01/1970") } ?? []
         self.transactionsFiltered = self.transactions
         self.linkedBrokerAccount = linkedBrokerAccount
-        self.numberOfDays = transactionHistoryResult.numberOfDaysHistory.intValue
+        self.numberOfDays = transactionHistoryResult.numberOfDaysHistory
     }
 
     func numberOfRows() -> Int {
